swtpm: Add support for --flag hlk-compliance
Add support for --flag hlk-compliance, if libtpms implements the newer
API call TPMLIB_SetFlags() that we then simply invoke and set the
TPMLIB_FLAG_HLK_COMPIANCE on libtpms.

The conditional support for of this feature allows us to still compile
with older versions of libtpms and doesn't force a new release of
libtpms that would make the API available.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>